refactor(settings): move miniapp  settings to a new page
