+++ src/renderer/src/pages/settings/DisplaySettings/DisplaySettings.tsx
@@ -1,8 +1,6 @@
 import { SyncOutlined } from '@ant-design/icons'
 import { isMac } from '@renderer/config/constant'
-import { DEFAULT_MIN_APPS } from '@renderer/config/minapps'
 import { useTheme } from '@renderer/context/ThemeProvider'
-import { useMinapps } from '@renderer/hooks/useMinapps'
 import { useSettings } from '@renderer/hooks/useSettings'
 import { useAppDispatch } from '@renderer/store'
 import {
@@ -19,7 +17,6 @@ import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
 import { SettingContainer, SettingDivider, SettingGroup, SettingRow, SettingRowTitle, SettingTitle } from '..'
-import MiniAppIconsManager from './MiniAppIconsManager'
 import SidebarIconsManager from './SidebarIconsManager'
 
 const DisplaySettings: FC = () => {
@@ -37,17 +34,13 @@ const DisplaySettings: FC = () => {
     showAssistantIcon,
     setShowAssistantIcon
   } = useSettings()
-  const { minapps, disabled, updateMinapps, updateDisabledMinapps } = useMinapps()
   const { theme: themeMode } = useTheme()
   const { t } = useTranslation()
   const dispatch = useAppDispatch()
 
   const [visibleIcons, setVisibleIcons] = useState(sidebarIcons?.visible || DEFAULT_SIDEBAR_ICONS)
   const [disabledIcons, setDisabledIcons] = useState(sidebarIcons?.disabled || [])
-  const [visibleMiniApps, setVisibleMiniApps] = useState(minapps)
-  const [disabledMiniApps, setDisabledMiniApps] = useState(disabled || [])
 
-  // 使用useCallback优化回调函数
   const handleWindowStyleChange = useCallback(
     (checked: boolean) => {
       setWindowStyle(checked ? 'transparent' : 'opaque')
@@ -60,13 +53,6 @@ const DisplaySettings: FC = () => {
     setDisabledIcons([])
     dispatch(setSidebarIcons({ visible: DEFAULT_SIDEBAR_ICONS, disabled: [] }))
   }, [dispatch])
-
-  const handleResetMinApps = useCallback(() => {
-    setVisibleMiniApps(DEFAULT_MIN_APPS)
-    setDisabledMiniApps([])
-    updateMinapps(DEFAULT_MIN_APPS)
-    updateDisabledMinapps([])
-  }, [updateDisabledMinapps, updateMinapps])
 
   const themeOptions = useMemo(
     () => [
@@ -175,22 +161,6 @@ const DisplaySettings: FC = () => {
           disabledIcons={disabledIcons}
           setVisibleIcons={setVisibleIcons}
           setDisabledIcons={setDisabledIcons}
-        />
-      </SettingGroup>
-      <SettingGroup theme={theme}>
-        <SettingTitle
-          style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between', alignItems: 'center' }}>
-          <span>{t('settings.display.minApp.title')}</span>
-          <ResetButtonWrapper>
-            <Button onClick={handleResetMinApps}>{t('common.reset')}</Button>
-          </ResetButtonWrapper>
-        </SettingTitle>
-        <SettingDivider />
-        <MiniAppIconsManager
-          visibleMiniApps={visibleMiniApps}
-          disabledMiniApps={disabledMiniApps}
-          setVisibleMiniApps={setVisibleMiniApps}
-          setDisabledMiniApps={setDisabledMiniApps}
         />
       </SettingGroup>
       <SettingGroup theme={theme}>
